+++ crates/maos-core/src/messages/notification.rs
@@ -160,3 +160,167 @@ impl Default for NotificationUrgency {
         Self::Normal
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::SessionId;
+    use serde_json::json;
+    use std::str::FromStr;
+
+    #[test]
+    fn test_notification_message_serialization() {
+        let notification = NotificationMessage {
+            message: "Build completed successfully".to_string(),
+            notification_type: NotificationType::TaskCompletion,
+            engineer_name: Some("Marvin".to_string()),
+            session_id: Some(
+                SessionId::from_str("sess_12345678-1234-1234-1234-123456789012").unwrap(),
+            ),
+            urgency: NotificationUrgency::Normal,
+            timestamp: chrono::Utc::now(),
+        };
+
+        // Should serialize to JSON
+        let json = serde_json::to_value(&notification).unwrap();
+        assert_eq!(json["message"], "Build completed successfully");
+        assert_eq!(json["notification_type"], "task_completion");
+        assert_eq!(json["urgency"], "normal");
+        assert_eq!(json["engineer_name"], "Marvin");
+
+        // Should deserialize from JSON
+        let deserialized: NotificationMessage = serde_json::from_value(json).unwrap();
+        assert_eq!(deserialized.message, notification.message);
+    }
+
+    #[test]
+    fn test_notification_types() {
+        assert_eq!(
+            serde_json::to_value(NotificationType::UserInputRequest).unwrap(),
+            json!("user_input_request")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationType::TaskCompletion).unwrap(),
+            json!("task_completion")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationType::AgentSpawned).unwrap(),
+            json!("agent_spawned")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationType::AgentCompleted).unwrap(),
+            json!("agent_completed")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationType::SecurityAlert).unwrap(),
+            json!("security_alert")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationType::SystemError).unwrap(),
+            json!("system_error")
+        );
+    }
+
+    #[test]
+    fn test_notification_urgency_levels() {
+        assert_eq!(
+            serde_json::to_value(NotificationUrgency::Low).unwrap(),
+            json!("low")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationUrgency::Normal).unwrap(),
+            json!("normal")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationUrgency::High).unwrap(),
+            json!("high")
+        );
+        assert_eq!(
+            serde_json::to_value(NotificationUrgency::Critical).unwrap(),
+            json!("critical")
+        );
+    }
+
+    #[test]
+    fn test_tts_formatting_with_engineer() {
+        let notification = NotificationMessage {
+            message: "tests passed".to_string(),
+            notification_type: NotificationType::TaskCompletion,
+            engineer_name: Some("Marvin".to_string()),
+            session_id: None,
+            urgency: NotificationUrgency::Normal,
+            timestamp: chrono::Utc::now(),
+        };
+
+        let tts = notification.to_tts_string();
+        assert_eq!(tts, "Marvin, task completed: tests passed");
+    }
+
+    #[test]
+    fn test_tts_formatting_without_engineer() {
+        let notification = NotificationMessage {
+            message: "need your input on PR review".to_string(),
+            notification_type: NotificationType::UserInputRequest,
+            engineer_name: None,
+            session_id: None,
+            urgency: NotificationUrgency::High,
+            timestamp: chrono::Utc::now(),
+        };
+
+        let tts = notification.to_tts_string();
+        assert_eq!(
+            tts,
+            "Engineer, I need your input: need your input on PR review"
+        );
+    }
+
+    #[test]
+    fn test_tts_formatting_all_types() {
+        // Test each notification type
+        let types_and_expected = vec![
+            (
+                NotificationType::UserInputRequest,
+                "test",
+                "Marv, I need your input: test",
+            ),
+            (
+                NotificationType::TaskCompletion,
+                "done",
+                "Marv, task completed: done",
+            ),
+            (
+                NotificationType::AgentSpawned,
+                "backend",
+                "New agent spawned: backend",
+            ),
+            (
+                NotificationType::AgentCompleted,
+                "frontend",
+                "Agent finished: frontend",
+            ),
+            (
+                NotificationType::SecurityAlert,
+                "danger",
+                "Security alert! danger",
+            ),
+            (
+                NotificationType::SystemError,
+                "crash",
+                "System error: crash",
+            ),
+        ];
+
+        for (notification_type, message, expected) in types_and_expected {
+            let notification = NotificationMessage {
+                message: message.to_string(),
+                notification_type,
+                engineer_name: Some("Marv".to_string()),
+                session_id: None,
+                urgency: NotificationUrgency::Normal,
+                timestamp: chrono::Utc::now(),
+            };
+
+            assert_eq!(notification.to_tts_string(), expected);
+        }
+    }
+}
