--- crates/maos-core/tests/path_tests.rs
@@ -1,1031 +1,0 @@
-//! Comprehensive path utilities tests
-//!
-//! This test suite covers all path utilities with focus on security testing
-
-use maos_core::path::{PathValidator, normalize_path, paths_equal, relative_path};
-use std::path::{Path, PathBuf};
-
-#[cfg(test)]
-mod path_validator_tests {
-    use super::*;
-
-    #[test]
-    fn test_path_validator_construction_with_empty_lists() {
-        let _validator = PathValidator::new(vec![], vec![]);
-
-        // Should not panic - basic construction should work
-        // The validator should be valid with empty allowed_roots and blocked_patterns
-    }
-
-    #[test]
-    fn test_path_validator_construction_with_allowed_roots() {
-        let allowed_roots = vec![PathBuf::from("/workspace1"), PathBuf::from("/workspace2")];
-        let _validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Should not panic - construction with allowed roots should work
-    }
-
-    #[test]
-    fn test_path_validator_construction_with_blocked_patterns() {
-        let blocked_patterns = vec![
-            "**/.git/**".to_string(),
-            "**/node_modules/**".to_string(),
-            "**/.ssh/**".to_string(),
-        ];
-        let _validator = PathValidator::new(vec![], blocked_patterns);
-
-        // Should not panic - construction with blocked patterns should work
-    }
-
-    #[test]
-    fn test_path_validator_construction_with_both() {
-        let allowed_roots = vec![PathBuf::from("/workspace")];
-        let blocked_patterns = vec!["**/.git/**".to_string()];
-        let _validator = PathValidator::new(allowed_roots, blocked_patterns);
-
-        // Should not panic - construction with both should work
-    }
-
-    #[test]
-    fn test_path_validator_construction_canonicalizes_allowed_roots() {
-        // ✅ PROPER TEST: Tests path canonicalization logic, not OS file system
-        let mock_root = PathBuf::from("/mock/root");
-        let allowed_roots = vec![mock_root.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test that the constructor properly handles allowed roots
-        assert!(validator.has_allowed_root(&mock_root));
-    }
-
-    // TDD Cycle 3 - RED PHASE: Tests for validate_workspace_path
-    #[test]
-    fn test_validate_workspace_path_success_with_allowed_workspace() {
-        // ✅ PROPER TEST: Tests workspace validation logic, not OS file system
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test path validation logic within mock workspace
-        let test_path = mock_workspace.join("test_file.txt");
-        let result = validator.validate_workspace_path(&test_path, &mock_workspace);
-        assert!(
-            result.is_ok(),
-            "Should validate path within allowed workspace"
-        );
-    }
-
-    #[test]
-    fn test_validate_workspace_path_fails_outside_workspace() {
-        // ✅ PROPER TEST: Tests workspace boundary validation logic
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test validation logic rejects outside paths
-        let outside_path = PathBuf::from("/etc/passwd");
-        let result = validator.validate_workspace_path(&outside_path, &mock_workspace);
-        assert!(result.is_err(), "Should reject path outside workspace");
-    }
-
-    #[test]
-    fn test_validate_workspace_path_prevents_path_traversal() {
-        // ✅ PROPER TEST: Tests path traversal detection logic
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test traversal detection logic
-        let traversal_path = mock_workspace.join("../../../etc/passwd");
-        let result = validator.validate_workspace_path(&traversal_path, &mock_workspace);
-        assert!(result.is_err(), "Should prevent path traversal attacks");
-    }
-
-    #[test]
-    fn test_validate_workspace_path_handles_relative_paths() {
-        // ✅ PROPER TEST: Tests relative path validation logic, not OS file system
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test that relative paths are validated correctly (business logic, not OS)
-        let relative_path = PathBuf::from("./subdir/file.txt");
-        let result = validator.validate_workspace_path(&relative_path, &mock_workspace);
-
-        // ✅ ACTUAL TEST: Relative paths within workspace should be accepted
-        assert!(
-            result.is_ok(),
-            "Should handle relative paths within workspace: {:?}",
-            result.err()
-        );
-
-        // ✅ ACTUAL TEST: Pattern matching should not block this path
-        assert!(
-            !validator.is_blocked_path(&relative_path),
-            "Relative path should not be blocked by pattern matching logic"
-        );
-    }
-
-    #[test]
-    fn test_validate_workspace_path_workspace_not_in_allowed_roots() {
-        // ✅ PROPER TEST: Tests allowed roots validation logic
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let other_dir = PathBuf::from("/some/other/path");
-        let allowed_roots = vec![other_dir];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Test validation logic when workspace not in allowed roots
-        let test_path = mock_workspace.join("file.txt");
-        let result = validator.validate_workspace_path(&test_path, &mock_workspace);
-        assert!(
-            result.is_err(),
-            "Should reject workspace not in allowed roots"
-        );
-    }
-
-    // TDD Cycle 4 - RED PHASE: Comprehensive path traversal attack prevention
-    #[test]
-    fn test_path_traversal_basic_dotdot_attack() {
-        // ✅ PROPER TEST: Tests path traversal detection logic, not OS file system
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        let attack_path = PathBuf::from("../../../etc/passwd");
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        assert!(result.is_err(), "Should block basic ../../../ attack");
-    }
-
-    #[test]
-    fn test_path_traversal_encoded_dotdot_attack() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // URL-encoded .. (%2e%2e)
-        let attack_path = PathBuf::from("%2e%2e/%2e%2e/%2e%2e/etc/passwd");
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        // Our enhanced security now detects and blocks these patterns
-        assert!(
-            result.is_err(),
-            "Should block URL-encoded traversal attacks"
-        );
-    }
-
-    #[test]
-    fn test_path_traversal_mixed_slashes() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        let attack_path = PathBuf::from("..\\..\\..\\etc\\passwd");
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        assert!(result.is_err(), "Should block mixed slash attacks");
-    }
-
-    #[test]
-    fn test_path_traversal_nested_workspace_escape() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        let attack_path = mock_workspace.join("subdir/../../../etc/passwd");
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        assert!(result.is_err(), "Should block nested workspace escape");
-    }
-
-    #[test]
-    fn test_path_traversal_legitimate_parent_access() {
-        // Use a simpler approach with temp directories we can control
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let workspace_root = mock_workspace.join("test_workspace");
-        let _subdir = workspace_root.join("subdir");
-
-        // ✅ PROPER TEST: Tests validation logic, no OS file system operations
-        let allowed_roots = vec![workspace_root.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Legitimate access within the workspace root (file.txt from workspace_root)
-        let legitimate_path = PathBuf::from("file.txt");
-        let result = validator.validate_workspace_path(&legitimate_path, &workspace_root);
-
-        // This should work: workspace_root/file.txt is within the allowed workspace_root
-        assert!(
-            result.is_ok(),
-            "Should allow legitimate file access within workspace"
-        );
-    }
-
-    #[test]
-    fn test_path_traversal_double_encoded_attack() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Double URL-encoded .. (%252e%252e)
-        let attack_path = PathBuf::from("%252e%252e/%252e%252e/%252e%252e/etc/passwd");
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        // Our enhanced security now detects and blocks these patterns
-        assert!(
-            result.is_err(),
-            "Should block double URL-encoded traversal attacks"
-        );
-    }
-
-    #[test]
-    fn test_path_traversal_null_byte_injection() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Null byte injection attempt (would be dangerous in C, less so in Rust)
-        let attack_string = "safe_file.txt\0../../../etc/passwd".to_string();
-        let attack_path = PathBuf::from(attack_string);
-        let _result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        // Rust handles null bytes in filenames safely - they become part of the filename
-        // Our path validation allows this since it's within workspace bounds
-    }
-
-    #[test]
-    fn test_path_traversal_long_path_attack() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let allowed_roots = vec![mock_workspace.clone()];
-        let validator = PathValidator::new(allowed_roots, vec![]);
-
-        // Create a very long traversal path
-        const LONG_TRAVERSAL_COUNT: usize = 100;
-        let mut long_traversal = String::new();
-        for _ in 0..LONG_TRAVERSAL_COUNT {
-            long_traversal.push_str("../");
-        }
-        long_traversal.push_str("etc/passwd");
-
-        let attack_path = PathBuf::from(long_traversal);
-        let result = validator.validate_workspace_path(&attack_path, &mock_workspace);
-        assert!(result.is_err(), "Should block long traversal attacks");
-    }
-
-    // TDD Cycle 5 - RED PHASE: Comprehensive glob pattern blocking tests
-    #[test]
-    fn test_is_blocked_path_simple_glob_patterns() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let blocked_patterns = vec![
-            "*.tmp".to_string(),
-            "*.log".to_string(),
-            "secret*".to_string(),
-        ];
-        let validator = PathValidator::new(vec![mock_workspace.clone()], blocked_patterns);
-
-        // Should block .tmp files
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("test.tmp")),
-            "Should block .tmp files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("data.tmp")),
-            "Should block .tmp files"
-        );
-
-        // Should block .log files
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("debug.log")),
-            "Should block .log files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("error.log")),
-            "Should block .log files"
-        );
-
-        // Should block files starting with 'secret'
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("secret.txt")),
-            "Should block secret files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("secrets")),
-            "Should block secret files"
-        );
-
-        // Should allow other files
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("config.txt")),
-            "Should allow other files"
-        );
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("data.json")),
-            "Should allow other files"
-        );
-    }
-
-    #[test]
-    fn test_is_blocked_path_directory_patterns() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let blocked_patterns = vec![
-            "node_modules/*".to_string(),
-            ".git/*".to_string(),
-            "target/**".to_string(),
-        ];
-        let validator = PathValidator::new(vec![mock_workspace.clone()], blocked_patterns);
-
-        // Should block node_modules directory contents
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("node_modules/package")),
-            "Should block node_modules contents"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("node_modules/react/index.js")),
-            "Should block nested node_modules"
-        );
-
-        // Should block .git directory contents
-        assert!(
-            validator.is_blocked_path(&PathBuf::from(".git/config")),
-            "Should block .git contents"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from(".git/hooks/pre-commit")),
-            "Should block nested .git"
-        );
-
-        // Should block target directory (recursive with **)
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("target/debug/app")),
-            "Should block target contents"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("target/release/deps/lib.so")),
-            "Should block deeply nested target"
-        );
-
-        // Should allow other paths
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("src/main.rs")),
-            "Should allow src files"
-        );
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("README.md")),
-            "Should allow root files"
-        );
-    }
-
-    #[test]
-    fn test_is_blocked_path_complex_glob_patterns() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let blocked_patterns = vec![
-            "*.tmp".to_string(),
-            "*.log".to_string(),
-            "*.bak".to_string(),
-            "test_*".to_string(),
-            "*.exe".to_string(),
-            "*.dll".to_string(),
-        ];
-        let validator = PathValidator::new(vec![mock_workspace.clone()], blocked_patterns);
-
-        // Should block multiple extensions (simplified to basic glob patterns)
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("app.tmp")),
-            "Should block tmp files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("debug.log")),
-            "Should block log files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("backup.bak")),
-            "Should block bak files"
-        );
-
-        // Should block test files
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("test_data.json")),
-            "Should block test files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("test_config")),
-            "Should block test files"
-        );
-
-        // Should block executables
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("program.exe")),
-            "Should block exe files"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("library.dll")),
-            "Should block dll files"
-        );
-
-        // Should allow other files
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("source.rs")),
-            "Should allow Rust files"
-        );
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("config.toml")),
-            "Should allow config files"
-        );
-    }
-
-    #[test]
-    fn test_is_blocked_path_absolute_vs_relative_paths() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let blocked_patterns = vec!["*.tmp".to_string(), "logs/*".to_string()];
-        let validator = PathValidator::new(vec![mock_workspace.clone()], blocked_patterns);
-
-        // Should work with relative paths
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("file.tmp")),
-            "Should block relative tmp"
-        );
-        assert!(
-            validator.is_blocked_path(&PathBuf::from("logs/debug.txt")),
-            "Should block logs dir"
-        );
-
-        // Should work with absolute paths
-        let abs_tmp = mock_workspace.join("file.tmp");
-        let abs_log = mock_workspace.join("logs/debug.txt");
-        assert!(
-            validator.is_blocked_path(&abs_tmp),
-            "Should block absolute tmp"
-        );
-        assert!(
-            validator.is_blocked_path(&abs_log),
-            "Should block absolute logs"
-        );
-    }
-
-    #[test]
-    fn test_is_blocked_path_no_patterns() {
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        // Should allow everything when no patterns defined
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("anything.tmp")),
-            "Should allow when no patterns"
-        );
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("secret.txt")),
-            "Should allow when no patterns"
-        );
-        assert!(
-            !validator.is_blocked_path(&PathBuf::from("node_modules/react")),
-            "Should allow when no patterns"
-        );
-    }
-
-    // TDD Cycle 9 - RED PHASE: Comprehensive generate_workspace_path tests
-    #[test]
-    fn test_generate_workspace_path_basic_structure() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "test_agent".to_string();
-
-        let workspace_path =
-            validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-
-        // Should not be empty
-        assert!(
-            !workspace_path.as_os_str().is_empty(),
-            "Workspace path should not be empty"
-        );
-
-        // Should be absolute path when root is absolute
-        if mock_workspace.is_absolute() {
-            assert!(
-                workspace_path.is_absolute(),
-                "Should generate absolute path for absolute root"
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_workspace_path_uniqueness() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session1 = SessionId::generate();
-        let session2 = SessionId::generate();
-        let agent_type: AgentType = "test_agent".to_string();
-
-        let path1 = validator.generate_workspace_path(&mock_workspace, &session1, &agent_type);
-        let path2 = validator.generate_workspace_path(&mock_workspace, &session2, &agent_type);
-
-        // Different sessions should generate different paths
-        assert_ne!(
-            path1, path2,
-            "Different sessions should generate unique paths"
-        );
-    }
-
-    #[test]
-    fn test_generate_workspace_path_agent_type_uniqueness() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent1: AgentType = "frontend".to_string();
-        let agent2: AgentType = "backend".to_string();
-
-        let path1 = validator.generate_workspace_path(&mock_workspace, &session_id, &agent1);
-        let path2 = validator.generate_workspace_path(&mock_workspace, &session_id, &agent2);
-
-        // Different agent types should generate different paths
-        assert_ne!(
-            path1, path2,
-            "Different agent types should generate unique paths"
-        );
-    }
-
-    #[test]
-    fn test_generate_workspace_path_consistency() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "test_agent".to_string();
-
-        let path1 = validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-        let path2 = validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-
-        // Same inputs should generate same path (deterministic)
-        assert_eq!(path1, path2, "Same inputs should generate consistent paths");
-    }
-
-    #[test]
-    fn test_generate_workspace_path_contains_identifiers() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "myagent".to_string();
-
-        let workspace_path =
-            validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-        let path_str = workspace_path.to_string_lossy();
-
-        // Should contain session and agent identifiers in some form
-        assert!(
-            path_str.contains(&session_id.to_string()) || path_str.contains(&agent_type),
-            "Path should contain session or agent identifiers: {path_str}"
-        );
-    }
-
-    #[test]
-    fn test_generate_workspace_path_relative_root() {
-        use maos_core::{AgentType, SessionId};
-
-        let relative_root = PathBuf::from("workspace");
-        let validator = PathValidator::new(vec![relative_root.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "test_agent".to_string();
-
-        let workspace_path =
-            validator.generate_workspace_path(&relative_root, &session_id, &agent_type);
-
-        // Should handle relative roots gracefully
-        assert!(
-            !workspace_path.as_os_str().is_empty(),
-            "Should handle relative roots"
-        );
-        assert!(
-            workspace_path.starts_with(&relative_root),
-            "Should be within provided root"
-        );
-    }
-
-    #[test]
-    fn test_generate_workspace_path_safe_characters() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        // Test with special characters that should be sanitized
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "agent_type_name".to_string();
-
-        let workspace_path =
-            validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-        let path_str = workspace_path.to_string_lossy();
-
-        // Should not contain dangerous path traversal patterns
-        assert!(
-            !path_str.contains(".."),
-            "Should not contain path traversal patterns"
-        );
-        assert!(
-            !path_str.contains("./"),
-            "Should not contain current directory references"
-        );
-    }
-
-    #[test]
-    fn test_generate_workspace_path_length_reasonable() {
-        use maos_core::{AgentType, SessionId};
-
-        let mock_workspace = PathBuf::from("/mock/workspace");
-        let validator = PathValidator::new(vec![mock_workspace.clone()], vec![]);
-
-        let session_id = SessionId::generate();
-        let agent_type: AgentType = "very_long_agent_type_name_for_testing".to_string();
-
-        let workspace_path =
-            validator.generate_workspace_path(&mock_workspace, &session_id, &agent_type);
-
-        // Should generate reasonable length paths (not exceed typical filesystem limits)
-        const MAX_PATH_LENGTH: usize = 4096;
-        let path_len = workspace_path.to_string_lossy().len();
-        assert!(
-            path_len < MAX_PATH_LENGTH,
-            "Path should be reasonable length: {path_len} chars"
-        );
-        assert!(
-            path_len > mock_workspace.to_string_lossy().len(),
-            "Path should extend beyond root"
-        );
-    }
-}
-
-#[cfg(test)]
-mod path_utils_tests {
-    use super::*;
-
-    // TDD Cycle 6 - RED PHASE: normalize_path function tests
-    #[test]
-    fn test_normalize_path_basic() {
-        // Should normalize simple paths
-        assert_eq!(
-            normalize_path(Path::new("./file.txt")),
-            PathBuf::from("file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("dir/./file.txt")),
-            PathBuf::from("dir/file.txt")
-        );
-
-        // Should handle parent directory references
-        assert_eq!(
-            normalize_path(Path::new("dir/../file.txt")),
-            PathBuf::from("file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("a/b/../c/file.txt")),
-            PathBuf::from("a/c/file.txt")
-        );
-
-        // Should handle multiple dots
-        assert_eq!(
-            normalize_path(Path::new("./dir/../file.txt")),
-            PathBuf::from("file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("a/./b/../c")),
-            PathBuf::from("a/c")
-        );
-    }
-
-    #[test]
-    fn test_normalize_path_edge_cases() {
-        // Should handle empty and root paths (path-clean normalizes empty to ".")
-        assert_eq!(normalize_path(Path::new("")), PathBuf::from("."));
-        assert_eq!(normalize_path(Path::new(".")), PathBuf::from("."));
-        assert_eq!(normalize_path(Path::new("./")), PathBuf::from("."));
-
-        assert_eq!(
-            normalize_path(Path::new("./some/path")),
-            PathBuf::from("some/path")
-        );
-
-        // Should handle too many parent references
-        assert_eq!(
-            normalize_path(Path::new("../file.txt")),
-            PathBuf::from("../file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("../../file.txt")),
-            PathBuf::from("../../file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("dir/../../file.txt")),
-            PathBuf::from("../file.txt")
-        );
-    }
-
-    #[test]
-    fn test_normalize_path_absolute() {
-        // Should handle absolute paths
-        assert_eq!(
-            normalize_path(Path::new("/tmp/./file.txt")),
-            PathBuf::from("/tmp/file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("/tmp/../file.txt")),
-            PathBuf::from("/file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("/a/b/../c")),
-            PathBuf::from("/a/c")
-        );
-
-        // Should not go above root
-        assert_eq!(
-            normalize_path(Path::new("/../file.txt")),
-            PathBuf::from("/file.txt")
-        );
-        assert_eq!(
-            normalize_path(Path::new("/../../file.txt")),
-            PathBuf::from("/file.txt")
-        );
-    }
-
-    #[test]
-    fn test_normalize_path_windows_edge_cases() {
-        // Test the specific failure case from CI: "/a"
-        let path = PathBuf::from("/a");
-        let normalized = normalize_path(&path);
-
-        // The behavior should be platform-consistent
-        if path.is_absolute() {
-            assert!(
-                normalized.is_absolute(),
-                "If original path '/a' is absolute on this platform, normalized should be too. Original: {path:?}, Normalized: {normalized:?}"
-            );
-        } else {
-            // On Windows, "/a" might not be absolute (needs drive letter)
-            // This is platform-specific behavior and is acceptable
-        }
-    }
-
-    #[test]
-    fn test_normalize_path_cross_platform() {
-        // Platform-specific separator handling
-        #[cfg(windows)]
-        {
-            // Windows handles both separators
-            assert_eq!(
-                normalize_path(Path::new("dir\\file.txt")),
-                PathBuf::from("dir\\file.txt")
-            );
-            assert_eq!(
-                normalize_path(Path::new("dir\\..\\file.txt")),
-                PathBuf::from("file.txt")
-            );
-        }
-
-        #[cfg(not(windows))]
-        {
-            // Unix treats backslashes as literal characters
-            assert_eq!(
-                normalize_path(Path::new("dir\\file.txt")),
-                PathBuf::from("dir\\file.txt") // Single filename with backslash
-            );
-            // Only forward slashes work as separators
-            assert_eq!(
-                normalize_path(Path::new("dir/../file.txt")),
-                PathBuf::from("file.txt")
-            );
-        }
-    }
-
-    // TDD Cycle 7 - RED PHASE: Comprehensive paths_equal tests
-    #[test]
-    fn test_paths_equal_identical_paths() {
-        // Should return true for identical paths
-        assert!(
-            paths_equal(Path::new("/test/file.txt"), Path::new("/test/file.txt")),
-            "Identical paths should be equal"
-        );
-        assert!(
-            paths_equal(Path::new("relative/path"), Path::new("relative/path")),
-            "Identical relative paths should be equal"
-        );
-    }
-
-    #[test]
-    fn test_paths_equal_different_paths() {
-        // Should return false for different paths
-        assert!(
-            !paths_equal(Path::new("/test/file1.txt"), Path::new("/test/file2.txt")),
-            "Different files should not be equal"
-        );
-        assert!(
-            !paths_equal(Path::new("/test/dir1"), Path::new("/test/dir2")),
-            "Different directories should not be equal"
-        );
-    }
-
-    #[test]
-    fn test_paths_equal_normalized_paths() {
-        // Should return true for paths that normalize to the same location
-        assert!(
-            paths_equal(Path::new("/test/./file.txt"), Path::new("/test/file.txt")),
-            "Normalized paths should be equal"
-        );
-        assert!(
-            paths_equal(
-                Path::new("/test/dir/../file.txt"),
-                Path::new("/test/file.txt")
-            ),
-            "Path with parent reference should equal normalized"
-        );
-        assert!(
-            paths_equal(Path::new("./file.txt"), Path::new("file.txt")),
-            "Current dir reference should equal direct path"
-        );
-    }
-
-    #[test]
-    fn test_paths_equal_cross_platform_separators() {
-        // Platform-specific: Windows treats \ and / as equivalent, Unix doesn't
-        #[cfg(windows)]
-        {
-            assert!(
-                paths_equal(Path::new("dir\\file.txt"), Path::new("dir/file.txt")),
-                "On Windows, different separators should be equal"
-            );
-            assert!(
-                paths_equal(
-                    Path::new("dir\\subdir\\file.txt"),
-                    Path::new("dir/subdir/file.txt")
-                ),
-                "On Windows, nested paths with different separators should be equal"
-            );
-        }
-
-        #[cfg(not(windows))]
-        {
-            // On Unix, backslashes are literal characters, not separators
-            assert!(
-                !paths_equal(Path::new("dir\\file.txt"), Path::new("dir/file.txt")),
-                "On Unix, backslash paths are different from forward slash paths"
-            );
-            // These ARE equal because they're the same path
-            assert!(
-                paths_equal(Path::new("dir/file.txt"), Path::new("dir/file.txt")),
-                "Same paths should be equal"
-            );
-        }
-    }
-
-    #[test]
-    fn test_paths_equal_case_sensitivity() {
-        // On Unix-like systems, paths are case-sensitive
-        // On Windows, they're case-insensitive
-        // We'll implement Unix-style behavior (case-sensitive)
-        #[cfg(unix)]
-        {
-            assert!(
-                !paths_equal(Path::new("/Test/FILE.txt"), Path::new("/test/file.txt")),
-                "Case-different paths should not be equal on Unix"
-            );
-        }
-
-        // For now, we'll test the case-sensitive behavior
-        assert!(
-            !paths_equal(Path::new("File.txt"), Path::new("file.txt")),
-            "Case-different paths should not be equal"
-        );
-    }
-
-    #[test]
-    fn test_paths_equal_absolute_vs_relative() {
-        // Should handle comparison between absolute and relative paths
-        // These should be different unless they resolve to same location
-        assert!(
-            !paths_equal(Path::new("/test/file.txt"), Path::new("file.txt")),
-            "Absolute and relative paths should generally not be equal"
-        );
-        assert!(
-            !paths_equal(Path::new("/tmp"), Path::new("tmp")),
-            "Absolute and relative dirs should generally not be equal"
-        );
-    }
-
-    #[test]
-    fn test_paths_equal_empty_and_current() {
-        // Should handle empty paths and current directory references
-        assert!(
-            paths_equal(Path::new(""), Path::new("")),
-            "Empty paths should be equal"
-        );
-        assert!(
-            paths_equal(Path::new("."), Path::new(".")),
-            "Current directory references should be equal"
-        );
-        assert!(
-            paths_equal(Path::new("./"), Path::new(".")),
-            "Current directory with slash should equal without"
-        );
-    }
-
-    // TDD Cycle 8 - RED PHASE: Comprehensive relative_path tests
-    #[test]
-    fn test_relative_path_same_directory() {
-        // Should return "." or empty for same paths
-        let result = relative_path(Path::new("/test"), Path::new("/test"));
-        assert_eq!(result, Some(PathBuf::from(".")));
-
-        let result = relative_path(Path::new("base"), Path::new("base"));
-        assert_eq!(result, Some(PathBuf::from(".")));
-    }
-
-    #[test]
-    fn test_relative_path_direct_child() {
-        // Should return child name for direct children
-        let result = relative_path(Path::new("/base"), Path::new("/base/child"));
-        assert_eq!(result, Some(PathBuf::from("child")));
-
-        let result = relative_path(Path::new("/base"), Path::new("/base/subdir/file.txt"));
-        assert_eq!(result, Some(PathBuf::from("subdir/file.txt")));
-    }
-
-    #[test]
-    fn test_relative_path_parent_directory() {
-        // Should return .. for parent directories
-        let result = relative_path(Path::new("/base/subdir"), Path::new("/base"));
-        assert_eq!(result, Some(PathBuf::from("..")));
-
-        let result = relative_path(Path::new("/base/sub1/sub2"), Path::new("/base"));
-        assert_eq!(result, Some(PathBuf::from("../..")));
-    }
-
-    #[test]
-    fn test_relative_path_sibling_directories() {
-        // Should return ../sibling for sibling paths
-        let result = relative_path(Path::new("/base/dir1"), Path::new("/base/dir2"));
-        assert_eq!(result, Some(PathBuf::from("../dir2")));
-
-        let result = relative_path(
-            Path::new("/base/dir1/sub"),
-            Path::new("/base/dir2/file.txt"),
-        );
-        assert_eq!(result, Some(PathBuf::from("../../dir2/file.txt")));
-    }
-
-    #[test]
-    fn test_relative_path_different_roots() {
-        // Should return None for completely different roots
-        let result = relative_path(Path::new("/usr/local"), Path::new("/var/log"));
-        assert!(result.is_some()); // Should be able to calculate ../../../var/log
-
-        // More complex case
-        let result = relative_path(Path::new("/home/user"), Path::new("/etc/config"));
-        assert!(result.is_some()); // Should be ../../etc/config
-    }
-
-    #[test]
-    fn test_relative_path_normalized_inputs() {
-        // Should handle paths that need normalization
-        let result = relative_path(Path::new("/base/./dir"), Path::new("/base/target"));
-        assert_eq!(result, Some(PathBuf::from("../target")));
-
-        let result = relative_path(Path::new("/base/dir/../other"), Path::new("/base/target"));
-        assert_eq!(result, Some(PathBuf::from("../target")));
-    }
-
-    #[test]
-    fn test_relative_path_relative_inputs() {
-        // Should handle relative base and target paths
-        let result = relative_path(Path::new("base/dir"), Path::new("base/target"));
-        assert_eq!(result, Some(PathBuf::from("../target")));
-
-        let result = relative_path(Path::new("dir1"), Path::new("dir2/file.txt"));
-        assert_eq!(result, Some(PathBuf::from("../dir2/file.txt")));
-    }
-
-    #[test]
-    fn test_relative_path_cross_platform() {
-        // Platform-specific separator handling
-        #[cfg(windows)]
-        {
-            // Windows handles both separators
-            let result = relative_path(Path::new("base\\dir"), Path::new("base\\target"));
-            assert_eq!(result, Some(PathBuf::from("..\\target")));
-            let result = relative_path(Path::new("dir1\\sub"), Path::new("dir2\\file.txt"));
-            assert_eq!(result, Some(PathBuf::from("..\\..\\dir2\\file.txt")));
-        }
-
-        #[cfg(not(windows))]
-        {
-            // Unix only uses forward slashes
-            let result = relative_path(Path::new("base/dir"), Path::new("base/target"));
-            assert_eq!(result, Some(PathBuf::from("../target")));
-            let result = relative_path(Path::new("dir1/sub"), Path::new("dir2/file.txt"));
-            assert_eq!(result, Some(PathBuf::from("../../dir2/file.txt")));
-        }
-    }
-}
