+++ crates/maos/src/cli/dispatcher.rs
@@ -481,4 +481,293 @@ mod tests {
             panic!("Expected ResourceLimit error, got: {result:?}");
         }
     }
+
+    #[tokio::test]
+    async fn test_handler_validation_failure() {
+        // Test that validation errors are properly propagated
+        struct ValidatingHandler;
+
+        #[async_trait]
+        impl CommandHandler for ValidatingHandler {
+            async fn execute(&self, _input: HookInput) -> Result<CommandResult> {
+                Ok(CommandResult {
+                    exit_code: ExitCode::Success,
+                    output: None,
+                    metrics: ExecutionMetrics::default(),
+                })
+            }
+
+            fn validate_input(&self, input: &HookInput) -> Result<()> {
+                if input.session_id.is_empty() {
+                    return Err(MaosError::InvalidInput {
+                        message: "Session ID cannot be empty".to_string(),
+                    });
+                }
+                Ok(())
+            }
+
+            fn name(&self) -> &'static str {
+                "validating_handler"
+            }
+        }
+
+        let config = Arc::new(MaosConfig::default());
+        let metrics = Arc::new(PerformanceMetrics::new());
+
+        let mock_input = HookInput {
+            session_id: "".to_string(), // Empty session ID should fail validation
+            transcript_path: "/tmp/test.jsonl".into(),
+            cwd: "/tmp".into(),
+            hook_event_name: maos_core::hook_constants::PRE_TOOL_USE.to_string(),
+            ..Default::default()
+        };
+
+        let input_provider = Box::new(MockInputProvider {
+            input: mock_input,
+            should_fail: false,
+        });
+
+        let dispatcher =
+            CommandDispatcher::new_with_input_provider(config, metrics, input_provider)
+                .await
+                .unwrap();
+
+        dispatcher.registry.register(
+            maos_core::hook_constants::PRE_TOOL_USE.to_string(),
+            Box::new(ValidatingHandler),
+        );
+
+        let command = Commands::PreToolUse;
+        let result = dispatcher.dispatch(command).await;
+
+        assert!(result.is_err());
+        if let Err(MaosError::InvalidInput { message }) = result {
+            assert_eq!(message, "Session ID cannot be empty");
+        } else {
+            panic!("Expected InvalidInput error");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_concurrent_dispatch_safety() {
+        // Test that multiple concurrent dispatches work correctly
+        let config = Arc::new(MaosConfig::default());
+        let metrics = Arc::new(PerformanceMetrics::new());
+
+        // Use mock input provider to avoid blocking on stdin
+        let mock_input = HookInput {
+            session_id: "test".to_string(),
+            transcript_path: "/tmp/test.jsonl".into(),
+            cwd: "/tmp".into(),
+            hook_event_name: maos_core::hook_constants::PRE_TOOL_USE.to_string(),
+            ..Default::default()
+        };
+
+        let input_provider = Box::new(MockInputProvider {
+            input: mock_input,
+            should_fail: false,
+        });
+
+        let dispatcher =
+            CommandDispatcher::new_with_input_provider(config, metrics, input_provider)
+                .await
+                .unwrap();
+
+        // Register simple handlers
+        for event in &[
+            maos_core::hook_constants::PRE_TOOL_USE,
+            maos_core::hook_constants::POST_TOOL_USE,
+            maos_core::hook_constants::NOTIFICATION,
+        ] {
+            dispatcher.registry.register(
+                event.to_string(),
+                Box::new(TestHandler {
+                    name: event,
+                    exit_code: ExitCode::Success,
+                }),
+            );
+        }
+
+        // Launch multiple concurrent dispatches
+        let dispatcher = Arc::new(dispatcher);
+        let mut handles = vec![];
+
+        for i in 0..10 {
+            let dispatcher = dispatcher.clone();
+            let handle = tokio::spawn(async move {
+                // Use different commands in rotation
+                let command = match i % 3 {
+                    0 => Commands::PreToolUse,
+                    1 => Commands::PostToolUse,
+                    _ => Commands::Notify,
+                };
+
+                // Commands that don't expect stdin should work without input
+                if !command.expects_stdin() {
+                    dispatcher.dispatch(command).await
+                } else {
+                    // For commands expecting stdin, they should fail gracefully
+                    dispatcher.dispatch(command).await
+                }
+            });
+            handles.push(handle);
+        }
+
+        // All dispatches should complete without panicking
+        for handle in handles {
+            let _ = handle.await.unwrap(); // Result can be Ok or Err, but shouldn't panic
+        }
+    }
+
+    #[tokio::test]
+    async fn test_memory_limit_warning() {
+        // Test that high memory usage triggers warning (doesn't fail execution)
+        struct MemoryIntensiveHandler;
+
+        #[async_trait]
+        impl CommandHandler for MemoryIntensiveHandler {
+            async fn execute(&self, _input: HookInput) -> Result<CommandResult> {
+                // Allocate a large vector to simulate memory usage
+                let _large_vec: Vec<u8> = vec![0; 50 * 1024 * 1024]; // 50MB
+
+                Ok(CommandResult {
+                    exit_code: ExitCode::Success,
+                    output: Some("Memory intensive operation completed".to_string()),
+                    metrics: ExecutionMetrics::default(),
+                })
+            }
+
+            fn name(&self) -> &'static str {
+                "memory_intensive"
+            }
+        }
+
+        let mut config = MaosConfig::default();
+        config.hooks.max_input_size_mb = 10; // Set low memory limit for testing
+        let config = Arc::new(config);
+        let metrics = Arc::new(PerformanceMetrics::new());
+
+        let mock_input = HookInput {
+            session_id: "test_session".to_string(),
+            transcript_path: "/tmp/test.jsonl".into(),
+            cwd: "/tmp".into(),
+            hook_event_name: maos_core::hook_constants::SESSION_START.to_string(),
+            ..Default::default()
+        };
+
+        let input_provider = Box::new(MockInputProvider {
+            input: mock_input,
+            should_fail: false,
+        });
+
+        let dispatcher =
+            CommandDispatcher::new_with_input_provider(config, metrics, input_provider)
+                .await
+                .unwrap();
+
+        dispatcher.registry.register(
+            maos_core::hook_constants::SESSION_START.to_string(),
+            Box::new(MemoryIntensiveHandler),
+        );
+
+        // Should succeed despite high memory usage (only warns)
+        let command = Commands::SessionStart;
+        let result = dispatcher.dispatch(command).await.unwrap();
+        assert_eq!(result.exit_code, ExitCode::Success);
+    }
+
+    #[tokio::test]
+    async fn test_zero_timeout_edge_case() {
+        // Test edge case with zero timeout
+        let mut config = MaosConfig::default();
+        config.system.max_execution_time_ms = 0; // Zero timeout
+        let config = Arc::new(config);
+        let metrics = Arc::new(PerformanceMetrics::new());
+
+        let mock_input = HookInput {
+            session_id: "test".to_string(),
+            transcript_path: "/tmp/test.jsonl".into(),
+            cwd: "/tmp".into(),
+            hook_event_name: maos_core::hook_constants::SESSION_START.to_string(),
+            ..Default::default()
+        };
+
+        let input_provider = Box::new(MockInputProvider {
+            input: mock_input,
+            should_fail: false,
+        });
+
+        let dispatcher =
+            CommandDispatcher::new_with_input_provider(config, metrics, input_provider)
+                .await
+                .unwrap();
+
+        dispatcher.registry.register(
+            maos_core::hook_constants::SESSION_START.to_string(),
+            Box::new(TestHandler {
+                name: "instant_handler",
+                exit_code: ExitCode::Success,
+            }),
+        );
+
+        // With 0ms timeout, the result depends on how fast the handler executes
+        // It might succeed if the handler is instant, or fail if timeout is checked first
+        let command = Commands::SessionStart;
+        let result = dispatcher.dispatch(command).await;
+        // Just verify it doesn't panic - the result can be Ok or Err
+        assert!(result.is_ok() || result.is_err());
+    }
+
+    #[tokio::test]
+    async fn test_handler_panic_recovery() {
+        // Test that panicking handlers don't crash the dispatcher
+        struct PanickingHandler;
+
+        #[async_trait]
+        impl CommandHandler for PanickingHandler {
+            async fn execute(&self, _input: HookInput) -> Result<CommandResult> {
+                panic!("Handler panic for testing");
+            }
+
+            fn name(&self) -> &'static str {
+                "panicking_handler"
+            }
+        }
+
+        let config = Arc::new(MaosConfig::default());
+        let metrics = Arc::new(PerformanceMetrics::new());
+
+        let mock_input = HookInput {
+            session_id: "test".to_string(),
+            transcript_path: "/tmp/test.jsonl".into(),
+            cwd: "/tmp".into(),
+            hook_event_name: maos_core::hook_constants::SESSION_START.to_string(),
+            ..Default::default()
+        };
+
+        let input_provider = Box::new(MockInputProvider {
+            input: mock_input,
+            should_fail: false,
+        });
+
+        let dispatcher =
+            CommandDispatcher::new_with_input_provider(config, metrics, input_provider)
+                .await
+                .unwrap();
+
+        dispatcher.registry.register(
+            maos_core::hook_constants::SESSION_START.to_string(),
+            Box::new(PanickingHandler),
+        );
+
+        // The dispatch should handle the panic gracefully
+        let command = Commands::SessionStart;
+        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
+            let runtime = tokio::runtime::Runtime::new().unwrap();
+            runtime.block_on(async { dispatcher.dispatch(command).await })
+        }));
+
+        // The panic should be caught somewhere in the async runtime
+        assert!(result.is_err() || result.unwrap().is_err());
+    }
 }
